pyproject.toml
[project]
name = "python-capability-service"
version = "2.0.0"
description = "Stateless processing capabilities for image2url"
authors = [
    {name = "Abel", email = "abel@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "Pillow>=10.2.0",
    "boto3>=1.34.19",
    "aiohttp>=3.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
